+++ framework/model/src/main/java/is/codion/framework/model/DefaultEntityApplicationModel.java
@@ -40,7 +40,7 @@ import static java.util.Objects.requireNonNull;
  * @param <R> the {@link EntityEditor} type
  */
 public class DefaultEntityApplicationModel<M extends EntityModel<M, E, T, R>, E extends EntityEditModel<M, E, T, R>,
-				T extends EntityTableModel<M, E, T, R>, R extends EntityEditor<M, E, T, R>> implements EntityApplicationModel<M, E, T, R> {
+				T extends EntityTableModel<M, E, T, R>, R extends EntityEditor> implements EntityApplicationModel<M, E, T, R> {
 
 	private final EntityConnectionProvider connectionProvider;
 	private final DefaultEntityModels<M, E, T, R> models;
@@ -103,7 +103,7 @@ public class DefaultEntityApplicationModel<M extends EntityModel<M, E, T, R>, E 
 	}
 
 	private final class DefaultEntityModels<M extends EntityModel<M, E, T, R>, E extends EntityEditModel<M, E, T, R>,
-					T extends EntityTableModel<M, E, T, R>, R extends EntityEditor<M, E, T, R>> implements EntityModels<M, E, T, R> {
+					T extends EntityTableModel<M, E, T, R>, R extends EntityEditor> implements EntityModels<M, E, T, R> {
 
 		private final Collection<M> entityModels;