+++ framework/model/src/main/java/is/codion/framework/model/DefaultEntityEditModel.java
@@ -41,7 +41,7 @@ import static java.util.Objects.requireNonNull;
  * @param <R> the {@link EntityEditor} type
  */
 public class DefaultEntityEditModel<M extends EntityModel<M, E, T, R>, E extends EntityEditModel<M, E, T, R>,
-				T extends EntityTableModel<M, E, T, R>, R extends EntityEditor<M, E, T, R>> implements EntityEditModel<M, E, T, R> {
+				T extends EntityTableModel<M, E, T, R>, R extends EntityEditor> implements EntityEditModel<M, E, T, R> {
 
 	private final R editor;
 	private final DefaultPersistTasks tasks;