+++ framework/model/src/main/java/is/codion/framework/model/EntityEditModel.java
@@ -45,7 +45,7 @@ import java.util.Collection;
  * @see #editor()
  */
 public interface EntityEditModel<M extends EntityModel<M, E, T, R>, E extends EntityEditModel<M, E, T, R>,
-				T extends EntityTableModel<M, E, T, R>, R extends EntityEditor<M, E, T, R>> {
+				T extends EntityTableModel<M, E, T, R>, R extends EntityEditor> {
 
 	/**
 	 * @return the type of the entity this edit model is based on