+++ framework/model/src/main/java/is/codion/framework/model/EntityModel.java
@@ -37,7 +37,7 @@ import java.util.Map;
  * @param <R> the {@link EntityEditor} type
  */
 public interface EntityModel<M extends EntityModel<M, E, T, R>, E extends EntityEditModel<M, E, T, R>,
-				T extends EntityTableModel<M, E, T, R>, R extends EntityEditor<M, E, T, R>> {
+				T extends EntityTableModel<M, E, T, R>, R extends EntityEditor> {
 
 	/**
 	 * @return the type of the entity this entity model is based on
@@ -103,7 +103,7 @@ public interface EntityModel<M extends EntityModel<M, E, T, R>, E extends Entity
 	 * @param <R> the {@link EntityEditor} type
 	 */
 	interface DetailModels<M extends EntityModel<M, E, T, R>, E extends EntityEditModel<M, E, T, R>,
-					T extends EntityTableModel<M, E, T, R>, R extends EntityEditor<M, E, T, R>> {
+					T extends EntityTableModel<M, E, T, R>, R extends EntityEditor> {
 
 		/**
 		 * @return an unmodifiable view of the detail models this model contains